requirements.txt
cloudant==2.15.0
Flask==1.1.2
flask_cors==3.0.10
joblib==1.1.0
paho_mqtt==1.6.1
pyserial==3.5
